pyproject.toml
[tool.poetry]
name = "easyDataverse"
version = "0.4.2"
description = "Lightweight Dataverse interface in Python to upload, download and update datasets found in Dataverse instances."
authors = ["Jan Range <jan.range@simtech.uni-stuttgart.de>"]
license = "MIT License"
readme = "Readme.md"
packages = [{ include = "easyDataverse" }]

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^2.7.1"
pydataverse = "^0.3.1"
pyaml = "^24.4.0"
xmltodict = "^0.13.0"
python-forge = "18.6.0"
anytree = "^2.12.1"
dotted-dict = "1.1.3"
rich = "^13.7.1"
nob = "^0.8.2"
nest-asyncio = "^1.6.0"
aiohttp = "^3.9.5"
aiodns = "^3.2.0"
dvuploader = "^0.2.3"
email-validator = "^2.1.1"

[tool.poetry.group.test.dependencies]
pytest-cov = "^5.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
